Extract terminal action handlers to dedicated module (Phase 4)
Created src/lib/terminal-actions.ts (127 lines) with 2 UI helper functions:
- handleRunNowClick() - Executes interval prompt for autonomous terminals
- startRename() - Handles inline terminal renaming

Both functions use dependency injection pattern for testability.

Updated main.ts to import these functions and pass dependencies:
- handleRunNowClick(id, { state })
- startRename(id, target, { state, saveCurrentConfig, render })

Impact:
- main.ts: Reduced from 1,983 → 1,904 lines (79 lines removed)
- Risk: Low (UI helpers with clear dependencies)
- Testing: Type checking passes ✅ Biome linting passes ✅

Cumulative Progress (Phases 1-4):
- Phase 1: -44 lines (pure utilities)
- Phase 2: -74 lines (event handlers)
- Phase 3: -134 lines (recovery handlers)
- Phase 4: -79 lines (terminal actions)
- Total removed from main.ts: 331 lines
- New focused modules: 4 files
